Census attribution + the nullable-flag adjudication (a fixer pass rejected)
Witnesses now carry their TEST (the census's own CONTEXT ThreadLocal,
set by both runners test-side — the first exec->lowering spelling
tripped Invariant 6d and was corrected). u_ord labeled BIGINT at its
CONSTRUCTION site (builder makes projection and label together —
label = wire truth; 20 rows -> agree).

ADJUDICATED (user challenge): OutputCol.nullable today MEANS
multiplicity-echo (PureSql.nullable is its only writer) — not a wire
contract; the engine has no such flag (nullability lives in pure
multiplicity; union pads and subtype markers are unlabeled plumbing
its Java assembly handles). The 6,472 rows rename to
bottom-mult-backlog: the wholesale re-label workload, not bugs. The
NullPadLabels fixer pass was built and DELETED same hour — a
builder+fixer split creating mixed-meaning flags; recorded in the T4
row so it stays dead. All nine gates green.

+++ core/src/main/java/com/legend/lowering/Fold.java
@@ -142,8 +142,10 @@ final class Fold {
             return outs;
         }
         var w = new java.util.ArrayList<>(outs);
+        // label = wire truth (typed-IR census witness u_ord := IntLit):
+        // the ordinal literal IS a BIGINT
         w.add(new com.legend.sql.OutputCol("u_ord",
-                com.legend.sql.SqlType.Scalar.INTEGER, false));
+                com.legend.sql.SqlType.Scalar.BIGINT, false));
         return w;
     }
